Build Local Geocoder

This constructs a local geocoder that is based on the latest release of the City of St. Louis's master address list. This function therefore requires an internet connection, and will error if your computer is offline. Since the actual geocoding is done with a second function, however, it is possible to build a geocoder and store it offline for repeated use.

gw_build_geocoder(class, crs = 4269, return = c("coords", "parcel", "zip"),
    include_units = FALSE)

Arguments

class
One of either "sf" or "tibble".

crs
A numeric code corresponding to the desired coordinate system for the column output if return includes "coords" as well as the object output if class is "sf".

return
Optional; A character scalar or vector that describes the type of information to be applied to the target data. Options include the City's address identification numbers (addrrecnum), parcel identification numbers (handle), zip-codes, and x and y coordinates (in decimal degrees).

include_units
A logical scalar; if TRUE, all individual records for apartment units will be included. If FALSE (default), only records for the overall building will be retained.

Value

A sf object or tibble with the requested data in the return argument as well as a address variable containing the full street address string.

See also