Add Areal Unit Identifiers to Point Data

Apply identifiers from a selection of commonly used areal units to point data stored in a sf object.

gw_identify(.data, to, side = "right", class = "sf")

Arguments

.data
A sf object

to
The string name of an areal unit to aggregate to: "block group", "tract", "precinct", "ward", or "neighborhood".

side
One of either "right" or "left" indicating where the identifier variable should be placed in the

class
One of either "sf" (default) or "tibble".

Value

gw_identify returns either a sf object or tibble with the requested data.