Standardized Street Suffixes

gw_suffix standardizes a given set of street suffix values to the USPS preferred suffix abbreviation.

gw_suffix(.data, suffix, overwrite = TRUE, newSuffix)

Arguments

.data
A tbl

suffix
A character vector within .data that contains street suffixes

overwrite
A logical scalar. Should the output overwrite the given variable?

newSuffix
A name for a new vector to be created if overwrite = FALSE

Value

gw_suffix returns a tibble with the requested output - either the suffix variable has been overwritten or a new variable with corrected suffix abbreviations has been added.

Note

gw_suffix requires that a number variable names be unused in the original data - suf_com, suf_cor, suf_pri and suf_std. If these names are present, gw_suffix will return an error. These variables are created temporarily as part of the matching process.